refactor: organize data files
